add angle calculation api in Math class

+++ packages/hetu_script/lib/utils/uid.dart
@@ -1,5 +1,6 @@
 import 'dart:math';
 
+/// Get a random UUID
 String randomUUID() {
   final output = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'
       .replaceAllMapped(RegExp('[xy]'), (match) {
@@ -10,14 +11,40 @@ String randomUUID() {
   return output;
 }
 
-String randomUID4([int? repeat]) {
-  repeat ??= 1;
-  final sb = StringBuffer();
-  for (var i = 0; i < repeat; ++i) {
-    sb.write(((Random().nextDouble() + 1) * 0x10000)
+/// Get a random id consists by number and letters.
+String randomUID([int? length]) {
+  length ??= 4;
+  assert(length >= 1);
+  if (length < 15) {
+    final id = ((Random().nextDouble() + 1) *
+            int.parse('0x1'.padRight(length + 3, '0')))
         .truncate()
         .toRadixString(16)
-        .substring(1));
+        .substring(1);
+    return id;
+  } else {
+    final output = StringBuffer();
+    final c = (length / 12).floor();
+    for (var i = 0; i < c; ++i) {
+      output.write(randomUID(12));
+    }
+    output.write(randomUID(length - c * 12));
+    return output.toString();
   }
-  return sb.toString();
+}
+
+/// Get a random id consists by number.
+String randomNID([int? length]) {
+  length ??= 8;
+  assert(length >= 1);
+  final r = Random(DateTime.now().millisecondsSinceEpoch);
+  final output = StringBuffer();
+  for (var i = 0; i < length; ++i) {
+    if (i == 0) {
+      output.write(r.nextInt(9) + 1);
+    } else {
+      output.write(r.nextInt(10));
+    }
+  }
+  return output.toString();
 }
